requirements.txt
atlassian_python_api==1.13.28
colorama==0.4.1
win_unicode_console==0.5
